# perf: encoder reduce allocs